Optimized Pipeline-Delete to use HTTP-GET (not DELETE because it's not enabled by default), added Confirm-Dialog, changed color of Delete-Button.

+++ ToSic.Eav.ManagementUI/EAV/PipelineDesigner/PipelineManagement.js
@@ -24,6 +24,9 @@ angular.module('pipelineManagement', ['pipelineManagementFactory']).
 
 		// Delete a Pipeline
 		$scope.delete = function (pipeline) {
+			if (!confirm('Delete Pipeline "' + pipeline.Name + ' (' + pipeline.EntityId + ')"?'))
+				return;
+
 			pipelineManagementFactory.deletePipeline(pipeline.EntityId).then(function () {
 				$scope.refresh();
 			}, function (reason) {
@@ -72,7 +75,7 @@ angular.module('pipelineManagementFactory', ['ngResource', 'eavGlobalConfigurati
 				}
 			},
 			deletePipeline: function (id) {
-				return pipelineResource.delete({ action: 'DeletePipeline', appId: appId, id: id }).$promise;
+				return pipelineResource.get({ action: 'DeletePipeline', appId: appId, id: id }).$promise;
 			}
 		}
 	});
